dynamic derive() salt w/aes iv
Add iv param to derive() to use in salt construction to prevent pre-computing keys- thanks to github.com/six-two for suggesting. This algo change is auto-detected to allow already-encrypted data to be decrypted with this newer better version. If you wrote code that manually calls .derive(), you'll have to slightly update it to pass the encrypted payload's iv as a 2nd argument after the plain key.